Add remote store tests for kafka ingestion
