+++ server/src/main/java/org/opensearch/index/engine/InternalEngine.java
@@ -104,7 +104,7 @@ import org.opensearch.index.seqno.LocalCheckpointTracker;
 import org.opensearch.index.seqno.SeqNoStats;
 import org.opensearch.index.seqno.SequenceNumbers;
 import org.opensearch.index.shard.OpenSearchMergePolicy;
-import org.opensearch.index.translog.InternalEngineTranslogManager;
+import org.opensearch.index.translog.EngineTranslogManager;
 import org.opensearch.index.translog.InternalTranslogManager;
 import org.opensearch.index.translog.Translog;
 import org.opensearch.index.translog.TranslogCorruptedException;
@@ -156,7 +156,7 @@ public class InternalEngine extends Engine {
      */
     private volatile long lastDeleteVersionPruneTimeMSec;
 
-    protected final InternalEngineTranslogManager translogManager;
+    protected final EngineTranslogManager translogManager;
     protected final IndexWriter indexWriter;
     protected final LocalCheckpointTracker localCheckpointTracker;
     protected final AtomicLong maxUnsafeAutoIdTimestamp = new AtomicLong(-1);
@@ -248,7 +248,7 @@ public class InternalEngine extends Engine {
         ExternalReaderManager externalReaderManager = null;
         OpenSearchReaderManager internalReaderManager = null;
         EngineMergeScheduler scheduler = null;
-        InternalEngineTranslogManager translogManagerRef = null;
+        EngineTranslogManager translogManagerRef = null;
         boolean success = false;
         try {
             this.lastDeleteVersionPruneTimeMSec = engineConfig.getThreadPool().relativeTimeInMillis();
@@ -352,7 +352,7 @@ public class InternalEngine extends Engine {
         logger.trace("created new InternalEngine");
     }
 
-    protected InternalEngineTranslogManager createTranslogManager(
+    protected EngineTranslogManager createTranslogManager(
         String translogUUID,
         TranslogDeletionPolicy translogDeletionPolicy,
         CompositeTranslogEventListener translogEventListener
